requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Deep Learning & NLP
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99

# DNS and Domain Analysis
dnspython>=2.4.0
python-whois>=0.8.0

# Web Framework
streamlit>=1.25.0

# Graph Analysis
networkx>=3.1

# Email Processing
email-validator>=2.0.0

# Utilities
python-dateutil>=2.8.2
python-dotenv>=1.0.0
